Test code (Experimental)
This is just messing with Unity and figuring out how it works from the ground up. It's fine if you want to look at  the code but is not usable for the project as I'm not using any 2D physics and basing it purely on vector maths.
Its also no commented.